% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_gemeente_van_2016_naar_2019_voor_postcode}
\alias{grenswijziging_gemeente_van_2016_naar_2019_voor_postcode}
\title{Matrix omzetting postcode naar regio 2016-2019}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 355 rows and 4050 columns.
}
\usage{
data(grenswijziging_gemeente_van_2016_naar_2019_voor_postcode)
}
\description{
Een matrix voor het omzetten van indicatoren vanuit postcode uit jaar 2016 naar jaar 2019 voor regio gemeente
De omzetting gebeurt naar rato van het aantal adressen (PC6 en huisnummer).
Woonfunctie of toevoeging aan het huisnummer worden niet meegenomen.
De omzet-matrices zijn op gemeenteniveau.
}
\keyword{datasets}
