% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_gemeente_van_2017_naar_2024}
\alias{grenswijziging_gemeente_van_2017_naar_2024}
\title{Matrix grenswijzigingen 2017-2024}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 47 rows and 93 columns.
}
\usage{
data(grenswijziging_gemeente_van_2017_naar_2024)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2017 naar jaar 2024 voor regio gemeente
De omzetting gebeurt naar rato van het aantal adressen (PC6 en huisnummer).
Woonfunctie of toevoeging aan het huisnummer worden niet meegenomen.
De omzet-matrices zijn op gemeenteniveau.
}
\keyword{datasets}
