% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_gemeente_van_2018_naar_2023}
\alias{grenswijziging_gemeente_van_2018_naar_2023}
\title{Matrix grenswijzigingen 2018-2023}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 31 rows and 69 columns.
}
\usage{
data(grenswijziging_gemeente_van_2018_naar_2023)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2018 naar jaar 2023 voor regio gemeente
De omzetting gebeurt naar rato van het aantal adressen (PC6 en huisnummer).
Woonfunctie of toevoeging aan het huisnummer worden niet meegenomen.
De omzet-matrices zijn op gemeenteniveau.
}
\keyword{datasets}
