% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_gemeente_van_2019_naar_2020}
\alias{grenswijziging_gemeente_van_2019_naar_2020}
\title{Matrix grenswijzigingen 2019-2020}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 0 rows and 0 columns.
}
\usage{
data(grenswijziging_gemeente_van_2019_naar_2020)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2019 naar jaar 2020 voor regio gemeente
De omzetting gebeurt naar rato van het aantal adressen (PC6 en huisnummer).
Woonfunctie of toevoeging aan het huisnummer worden niet meegenomen.
De omzet-matrices zijn op gemeenteniveau.
}
\keyword{datasets}
