% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_gemeente_van_2022_naar_2023_voor_postcode}
\alias{grenswijziging_gemeente_van_2022_naar_2023_voor_postcode}
\title{Matrix omzetting postcode naar regio 2022-2023}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 342 rows and 4071 columns.
}
\usage{
data(grenswijziging_gemeente_van_2022_naar_2023_voor_postcode)
}
\description{
Een matrix voor het omzetten van indicatoren vanuit postcode uit jaar 2022 naar jaar 2023 voor regio gemeente
De omzetting gebeurt naar rato van het aantal adressen (PC6 en huisnummer).
Woonfunctie of toevoeging aan het huisnummer worden niet meegenomen.
De omzet-matrices zijn op gemeenteniveau.
}
\keyword{datasets}
