% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_toevoeging_gemeente_van_2016_naar_2017}
\alias{grenswijziging_toevoeging_gemeente_van_2016_naar_2017}
\title{Matrix grenswijzigingen 2016-2017}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 17 rows and 19 columns.
}
\usage{
data(grenswijziging_toevoeging_gemeente_van_2016_naar_2017)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2016 naar jaar 2017 voor regio gemeente.
De omzetting gebeurt naar rato van het aantal volledige adressen met
woonfunctie (PC6 en huisnummer, huisnummertoevoeging).
De omzet-matrices zijn op gemeenteniveau.
}
\keyword{datasets}
