% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_toevoeging_gemeente_van_2017_naar_2018}
\alias{grenswijziging_toevoeging_gemeente_van_2017_naar_2018}
\title{Matrix grenswijzigingen 2017-2018}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 16 rows and 24 columns.
}
\usage{
data(grenswijziging_toevoeging_gemeente_van_2017_naar_2018)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2017 naar jaar 2018 voor regio gemeente.
De omzetting gebeurt naar rato van het aantal volledige adressen met
woonfunctie (PC6 en huisnummer, huisnummertoevoeging).
De omzet-matrices zijn op gemeenteniveau.
}
\keyword{datasets}
