% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_toevoeging_gemeente_van_2018_naar_2021}
\alias{grenswijziging_toevoeging_gemeente_van_2018_naar_2021}
\title{Matrix grenswijzigingen 2018-2021}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 22 rows and 50 columns.
}
\usage{
data(grenswijziging_toevoeging_gemeente_van_2018_naar_2021)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2018 naar jaar 2021 voor regio gemeente.
De omzetting gebeurt naar rato van het aantal volledige adressen met
woonfunctie (PC6 en huisnummer, huisnummertoevoeging).
De omzet-matrices zijn op gemeenteniveau.
}
\keyword{datasets}
