% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_toevoeging_gemeente_van_2019_naar_2021}
\alias{grenswijziging_toevoeging_gemeente_van_2019_naar_2021}
\title{Matrix grenswijzigingen 2019-2021}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 7 rows and 10 columns.
}
\usage{
data(grenswijziging_toevoeging_gemeente_van_2019_naar_2021)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2019 naar jaar 2021 voor regio gemeente.
De omzetting gebeurt naar rato van het aantal volledige adressen met
woonfunctie (PC6 en huisnummer, huisnummertoevoeging).
De omzet-matrices zijn op gemeenteniveau.
}
\keyword{datasets}
