% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_toevoeging_gemeente_van_2020_naar_2022}
\alias{grenswijziging_toevoeging_gemeente_van_2020_naar_2022}
\title{Matrix grenswijzigingen 2020-2022}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 14 rows and 24 columns.
}
\usage{
data(grenswijziging_toevoeging_gemeente_van_2020_naar_2022)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2020 naar jaar 2022 voor regio gemeente.
De omzetting gebeurt naar rato van het aantal volledige adressen met
woonfunctie (PC6 en huisnummer, huisnummertoevoeging).
De omzet-matrices zijn op gemeenteniveau.
}
\keyword{datasets}
