% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_toevoeging_gemeente_van_2022_naar_2024}
\alias{grenswijziging_toevoeging_gemeente_van_2022_naar_2024}
\title{Matrix grenswijzigingen 2022-2024}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 7 columns.
}
\usage{
data(grenswijziging_toevoeging_gemeente_van_2022_naar_2024)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2022 naar jaar 2024 voor regio gemeente.
De omzetting gebeurt naar rato van het aantal volledige adressen met
woonfunctie (PC6 en huisnummer, huisnummertoevoeging).
De omzet-matrices zijn op gemeenteniveau.
}
\keyword{datasets}
