% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_toevoeging_wijk_van_2018_naar_2019}
\alias{grenswijziging_toevoeging_wijk_van_2018_naar_2019}
\title{Matrix grenswijzigingen 2018-2019}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 392 rows and 315 columns.
}
\usage{
data(grenswijziging_toevoeging_wijk_van_2018_naar_2019)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2018 naar jaar 2019 voor regio wijk.
De omzetting gebeurt naar rato van het aantal volledige adressen met
woonfunctie (PC6 en huisnummer, huisnummertoevoeging).
De omzet-matrices zijn op wijkniveau.
}
\keyword{datasets}
