% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_toevoeging_wijk_van_2019_naar_2022}
\alias{grenswijziging_toevoeging_wijk_van_2019_naar_2022}
\title{Matrix grenswijzigingen 2019-2022}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 615 rows and 449 columns.
}
\usage{
data(grenswijziging_toevoeging_wijk_van_2019_naar_2022)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2019 naar jaar 2022 voor regio wijk.
De omzetting gebeurt naar rato van het aantal volledige adressen met
woonfunctie (PC6 en huisnummer, huisnummertoevoeging).
De omzet-matrices zijn op wijkniveau.
}
\keyword{datasets}
