% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_toevoeging_wijk_van_2020_naar_2024}
\alias{grenswijziging_toevoeging_wijk_van_2020_naar_2024}
\title{Matrix grenswijzigingen 2020-2024}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 767 rows and 566 columns.
}
\usage{
data(grenswijziging_toevoeging_wijk_van_2020_naar_2024)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2020 naar jaar 2024 voor regio wijk.
De omzetting gebeurt naar rato van het aantal volledige adressen met
woonfunctie (PC6 en huisnummer, huisnummertoevoeging).
De omzet-matrices zijn op wijkniveau.
}
\keyword{datasets}
