% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_toevoeging_wijk_van_2022_naar_2023}
\alias{grenswijziging_toevoeging_wijk_van_2022_naar_2023}
\title{Matrix grenswijzigingen 2022-2023}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2293 rows and 2281 columns.
}
\usage{
data(grenswijziging_toevoeging_wijk_van_2022_naar_2023)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2022 naar jaar 2023 voor regio wijk.
De omzetting gebeurt naar rato van het aantal volledige adressen met
woonfunctie (PC6 en huisnummer, huisnummertoevoeging).
De omzet-matrices zijn op wijkniveau.
}
\keyword{datasets}
