% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_toevoeging_wijk_van_2023_naar_2024}
\alias{grenswijziging_toevoeging_wijk_van_2023_naar_2024}
\title{Matrix grenswijzigingen 2023-2024}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 124 rows and 83 columns.
}
\usage{
data(grenswijziging_toevoeging_wijk_van_2023_naar_2024)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2023 naar jaar 2024 voor regio wijk.
De omzetting gebeurt naar rato van het aantal volledige adressen met
woonfunctie (PC6 en huisnummer, huisnummertoevoeging).
De omzet-matrices zijn op wijkniveau.
}
\keyword{datasets}
