% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_wijk_van_2016_naar_2023}
\alias{grenswijziging_wijk_van_2016_naar_2023}
\title{Matrix grenswijzigingen 2016-2023}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1772 rows and 1399 columns.
}
\usage{
data(grenswijziging_wijk_van_2016_naar_2023)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2016 naar jaar 2023 voor regio wijk
De omzetting gebeurt naar rato van het aantal adressen (PC6 en huisnummer).
Woonfunctie of toevoeging aan het huisnummer worden niet meegenomen.
De omzet-matrices zijn op wijkniveau.
}
\keyword{datasets}
