% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_wijk_van_2017_naar_2020_voor_postcode}
\alias{grenswijziging_wijk_van_2017_naar_2020_voor_postcode}
\title{Matrix omzetting postcode naar regio 2017-2020}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3168 rows and 4061 columns.
}
\usage{
data(grenswijziging_wijk_van_2017_naar_2020_voor_postcode)
}
\description{
Een matrix voor het omzetten van indicatoren vanuit postcode uit jaar 2017 naar jaar 2020 voor regio wijk
De omzetting gebeurt naar rato van het aantal adressen (PC6 en huisnummer).
Woonfunctie of toevoeging aan het huisnummer worden niet meegenomen.
De omzet-matrices zijn op wijkniveau.
}
\keyword{datasets}
