% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_wijk_van_2018_naar_2021}
\alias{grenswijziging_wijk_van_2018_naar_2021}
\title{Matrix grenswijzigingen 2018-2021}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 833 rows and 671 columns.
}
\usage{
data(grenswijziging_wijk_van_2018_naar_2021)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2018 naar jaar 2021 voor regio wijk
De omzetting gebeurt naar rato van het aantal adressen (PC6 en huisnummer).
Woonfunctie of toevoeging aan het huisnummer worden niet meegenomen.
De omzet-matrices zijn op wijkniveau.
}
\keyword{datasets}
