% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_wijk_van_2019_naar_2024}
\alias{grenswijziging_wijk_van_2019_naar_2024}
\title{Matrix grenswijzigingen 2019-2024}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 966 rows and 740 columns.
}
\usage{
data(grenswijziging_wijk_van_2019_naar_2024)
}
\description{
Een matrix voor het omzetten van indicatoren uit jaar 2019 naar jaar 2024 voor regio wijk
De omzetting gebeurt naar rato van het aantal adressen (PC6 en huisnummer).
Woonfunctie of toevoeging aan het huisnummer worden niet meegenomen.
De omzet-matrices zijn op wijkniveau.
}
\keyword{datasets}
