% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenswijziging_wijk_van_2020_naar_2023_voor_postcode}
\alias{grenswijziging_wijk_van_2020_naar_2023_voor_postcode}
\title{Matrix omzetting postcode naar regio 2020-2023}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3345 rows and 4069 columns.
}
\usage{
data(grenswijziging_wijk_van_2020_naar_2023_voor_postcode)
}
\description{
Een matrix voor het omzetten van indicatoren vanuit postcode uit jaar 2020 naar jaar 2023 voor regio wijk
De omzetting gebeurt naar rato van het aantal adressen (PC6 en huisnummer).
Woonfunctie of toevoeging aan het huisnummer worden niet meegenomen.
De omzet-matrices zijn op wijkniveau.
}
\keyword{datasets}
